requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT == 2.4.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
webencodings==0.5.1
drf-yasg==1.21.3
django-ckeditor==6.4.2
whitenoise==6.2.0
gunicorn == 20.1.0
dj-database-url